requirements-dev.txt
-r requirements.txt
pytest>=8.0
pytest-asyncio>=0.24.0
pytest-httpx>=0.35.0
